+++ tarea1.h
@@ -10,6 +10,6 @@ int min(int val1, int val2);
 
 int compararTamano( char * str1, char * str2);
 
-long long obtenerValor(char *str1, char *str2);
+int obtenerValor(char *str1, char *str2);
 
 void completeMatrix(int **matrix, char *string1, char *string2);
